requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-user-agents==0.4.0
djangorestframework==3.10.3
djoser==2.0.3
gunicorn==19.9.0
psycopg2==2.8.4
ptyprocess==0.6.0
user-agents==2.0
whitenoise==4.1.4
